Refactorisation des tests.
Déplacement des méthodes de test de `MainForm.cs` vers une nouvelle classe `Tests` dans `Tests.cs` pour une meilleure organisation.

Les méthodes de test sont maintenant documentées avec des commentaires XML. Les appels aux anciennes méthodes dans `MainForm.cs` ont été remplacés par des appels aux nouvelles méthodes de la classe `Tests`, améliorant ainsi la lisibilité et la maintenabilité du code.
